Turns out I need two different cooldown implementations: one for space maps and one for planet. And that in turn means I need a lighter version of IMstaticSpaceObject: IStaticMapObject.
Introducing new design: if some model needs to be updated each second,
implement ISelfUpdating. Later there will be controllers that will call
appropriate methods to update the models. We will have green threads
(sort of) by that time. Views (components) will have to check flags
in change_flag namespace on models to see if they need to be redrawn.
Interface for the views will be introduced later. There is a great chance
that ISelfUpdating will be modified often in the near future too.